Restructure DisplayPatterns into subfolders
