+++ lib/DisplayPatternLib/src/DisplayPattern.cpp
