+++ lib/DisplayPatternLib/src/DisplayPatternLib.h
@@ -3,11 +3,6 @@
 
 // This header file is only in place to link to the other files in the common lib.
 
-#include "DisplayPatternLib.h"
 #include "PatternFactory.h"
-#include "CenterOutDisplayPattern.h"
-#include "RandomDisplayPattern.h"
-#include "SimpleShiftDisplayPattern.h"
-#include "SolidDisplayPattern.h"
 
 #endif
