+++ lib/DisplayPatternLib/src/DisplayPatterns/CenterOutDisplayPattern.cpp
