+++ lib/DisplayPatternLib/src/DisplayPatterns/CenterOutDisplayPattern.h
