+++ lib/DisplayPatternLib/src/DisplayPatterns/LowPowerDisplayPattern.cpp
