+++ lib/DisplayPatternLib/src/DisplayPatterns/LowPowerDisplayPattern.h
