+++ lib/DisplayPatternLib/src/DisplayPatterns/SimpleShiftDisplayPattern.cpp
