+++ lib/DisplayPatternLib/src/DisplayPatterns/SimpleShiftDisplayPattern.h
