+++ lib/DisplayPatternLib/src/DisplayPatterns/SolidDisplayPattern.cpp
