+++ lib/DisplayPatternLib/src/DisplayPatterns/SolidDisplayPattern.h
