+++ lib/DisplayPatternLib/src/PatternFactory.cpp
