+++ lib/DisplayPatternLib/src/PatternFactory.h
@@ -3,13 +3,13 @@
 
 #include <Arduino.h>
 #include "DisplayPattern.h"
-#include "SolidDisplayPattern.h"
-#include "SimpleShiftDisplayPattern.h"
-#include "RandomDisplayPattern.h"
-#include "CenterOutDisplayPattern.h"
-#include "LowPowerDisplayPattern.h"
+#include "DisplayPatterns\SolidDisplayPattern.h"
+#include "DisplayPatterns\SimpleShiftDisplayPattern.h"
+#include "DisplayPatterns\RandomDisplayPattern.h"
+#include "DisplayPatterns\CenterOutDisplayPattern.h"
+#include "DisplayPatterns\LowPowerDisplayPattern.h"
 #include <ColorPatternLib.h>
-#include "PixelBuffer.h"
+#include <PixelBuffer.h>
 #include <PatternData.h>
 #include <vector>
